refactor(push-notifications): extract legacy page and add feature flag routing
