+++ app/admin/push-notifications/push-notifications-v2.tsx
@@ -1,0 +1,5 @@
+'use client';
+
+export default function PushNotificationsV2() {
+  return <div>Push Notifications V2 (placeholder)</div>;
+}
